# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OMEN Signal Intelligence API",
    "description": "## Signal Intelligence Engine\n\nOMEN transforms market events into structured probability signals.\n\n### Authentication\n\nAll API endpoints require authentication via API key.\nInclude the `X-API-Key` header in all requests.\n\n**Public endpoints (no auth required):**\n- `/health/*` - Health checks\n- `/docs`, `/redoc` - API documentation\n- `/metrics` - Prometheus metrics\n\n### What OMEN provides\n- Probability assessment with confidence bounds\n- Signal validation and quality scoring\n- Geographic and temporal context\n- Full evidence chain and traceability\n\n### What OMEN does NOT provide\n- Impact assessment (severity, delay, cost)\n- Decision steering (temporal proximity, relevance rank, high-confidence flags)\n- Recommendations or advice\n- Risk quantification\n\nImpact assessment is the responsibility of downstream systems.",
    "version": "2.0.0",
    "x-logo": {
      "url": "https://omen.io/logo.png",
      "altText": "OMEN Logo"
    },
    "contact": {
      "name": "OMEN API Support",
      "url": "https://github.com/omen/omen",
      "email": "support@omen.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "paths": {
...
    "/api/v1/signals/{signal_id}/explanation": {
      "get": {
        "tags": [
          "Explanations"
        ],
        "summary": "Get Signal Explanation",
        "description": "Get detailed explanation for a signal.\n\nFormats:\n- json: Machine-readable audit format\n- text: Human-readable report",
        "operationId": "get_signal_explanation_api_v1_signals__signal_id__explanation_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Signal Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "json",
              "title": "Format"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parameters": {
      "get": {
        "tags": [
          "Explanations"
        ],
        "summary": "List All Parameters",
        "description": "List all rule parameters used by OMEN.\n\nUseful for:\n- Documentation\n- Auditing assumptions\n- Parameter sensitivity analysis",
        "operationId": "list_all_parameters_api_v1_parameters_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/activity": {
      "get": {
        "tags": [
          "Activity",
          "Activity"
        ],
        "summary": "Get Activity",
        "description": "Get recent activity from the OMEN system.\n\nAll activity comes from real pipeline execution and source fetches:\n- signal: When signals are generated\n- validation: When events pass or fail validation\n- rule: When translation rules are applied\n- alert: When high-confidence signals are created\n- source: When data is fetched from sources\n- error: When errors occur\n\nIf the pipeline has not run, the list is empty.",
        "operationId": "get_activity_api_v1_activity_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "description": "Maximum items to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum items to return"
          },
          {
            "name": "type_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by type: signal, validation, rule, alert, source, error, system",
              "title": "Type Filter"
            },
            "description": "Filter by type: signal, validation, rule, alert, source, error, system"
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityItemResponse"
                  },
                  "title": "Response Get Activity Api V1 Activity Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/activity/types": {
      "get": {
        "tags": [
          "Activity",
          "Activity"
        ],
        "summary": "Get Activity Types",
        "description": "Return available activity types and short descriptions.",
        "operationId": "get_activity_types_api_v1_activity_types_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Get Activity Types Api V1 Activity Types Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/debug/rejections": {
      "get": {
        "tags": [
          "Debug (DEV ONLY)",
          "Debug"
        ],
        "summary": "Get Rejections",
        "description": "Get recent rejected events with reasons.\n\nUse this to understand WHY events are being filtered out.\n\n**Requires scope:** `debug`",
        "operationId": "get_rejections_api_v1_debug_rejections_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ingestion",
                    "mapping",
                    "validation",
                    "translation",
                    "generation"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/debug/passed": {
      "get": {
        "tags": [
          "Debug (DEV ONLY)",
          "Debug"
        ],
        "summary": "Get Passed",
        "description": "Get recently passed/generated signals.\n\n**Requires scope:** `debug`",
        "operationId": "get_passed_api_v1_debug_passed_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/debug/statistics": {
      "get": {
        "tags": [
          "Debug (DEV ONLY)",
          "Debug"
        ],
        "summary": "Get Pipeline Statistics",
        "description": "Get overall pipeline statistics.\n\nShows pass/rejection rates and breakdown by stage.\n\n**Requires scope:** `debug`",
        "operationId": "get_pipeline_statistics_api_v1_debug_statistics_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/debug/data": {
      "delete": {
        "tags": [
          "Debug (DEV ONLY)",
          "Debug"
        ],
        "summary": "Clear Debug Data",
        "description": "Clear all debug records.\n\n**Requires scope:** `debug`",
        "operationId": "clear_debug_data_api_v1_debug_data_delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActivityItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "signal",
              "validation",
              "rule",
              "alert",
              "source",
              "error",
              "system"
            ],
            "title": "Type"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "message",
          "timestamp"
        ],
        "title": "ActivityItemResponse",
        "description": "Single activity item from the system."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token (optional, if enabled)"
      }
    }
  },
  "servers": [
    {
      "url": "https://api.omen.io",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.omen.io",
      "description": "Staging server"
...
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
}